fix variable data types

+++ apps/app-internal.h
@@ -49,9 +49,9 @@ void set_verbosity(const char *name, enum kcapi_verbosity level);
 void hex2bin(const char *hex, uint32_t hexlen, uint8_t *bin, uint32_t binlen);
 int hex2bin_alloc(const char *hex, uint32_t hexlen,
 		  uint8_t **bin, uint32_t *binlen);
-void bin2hex(const uint8_t *bin, uint32_t binlen,
-	     char *hex, uint32_t hexlen, int u);
-void bin2print(const uint8_t *bin, uint32_t binlen,
+void bin2hex(const uint8_t *bin, size_t binlen,
+	     char *hex, size_t hexlen, int u);
+void bin2print(const uint8_t *bin, size_t binlen,
 	       const char *filename, FILE *outfile, uint32_t lfcr);
 int read_complete(int fd, uint8_t *buf, uint32_t buflen);
 int check_filetype(int fd, struct stat *sb, const char *filename);
